Datos actualizados de varios php

+++ web/php/subirImagen.php
@@ -1,18 +1,23 @@
 <?php
 echo $_FILES["file"]["name"];
-if (($_FILES["file"]["type"] == "image/pjpeg")
-    || ($_FILES["file"]["type"] == "image/jpeg")
-    || ($_FILES["file"]["type"] == "image/png")
-    || ($_FILES["file"]["type"] == "image/gif")) {
-    if (move_uploaded_file($_FILES["file"]["tmp_name"], "../source/image/".$_FILES['file']['name'])) {
-        //more code here...
-        echo "../source/image/".$_FILES['file']['name'];
-    } else {
-        echo " No se ha movido la imagen";
-    }
 
-  
+
+if (($_FILES["file"]["type"] == "image/pjpeg") || ($_FILES["file"]["type"] == "image/jpeg")  || ($_FILES["file"]["type"] == "image/png")  || ($_FILES["file"]["type"] == "image/gif")) {
+  if (file_exists("../source/image/")) {
+    echo "Archivo ya Existente";
+  } else {
+    if (move_uploaded_file($_FILES["file"]["tmp_name"], "../source/image/" . $_FILES['file']['name'])) {
+      echo "../source/image/" . $_FILES['file']['name'];
+    } else {
+      echo " No se ha movido la imagen";
+    }
+  }
+} else {
+  echo "Tipo de archivo no soportado";
 }
+
+
+
 
 
 
@@ -73,4 +78,3 @@ if ($uploadOk == 0) {
 
 //var_dump("");
 */
-?>
